LPS-161481 Restructure to avoid potential StaleObjectStateException caused by AUTO flush mode of Hibernate
validateCompanyMaxUsers() may trigger the flushing of the user and updating its mvccVersion in entity cache which leads to StaleObjectStateException.
To avoid this, we have to ensure to have an up-to-date user by fetching it after the validateCompanyMaxUsers() call.
To support this, the condition around the validation can be simplified, because currently there are no use cases where we are approving an already approved user.